# postman/api.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: FinchAPI
servers:
  - url: "https://finchapp.eastus.cloudapp.azure.com/api"
  - url: "https://localhost:8888/api"
  - url: "https://localhost:8080/"
paths:
  /signup:
    post:
      summary: "Sign up a user"
      operationId: signupUser
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupPayload"
      responses:
        "200":
          description: "User Info"
          headers:
            Set-Cookie:
              description: >
                Contains the session cookies named `sid` and `user_id`.
                Pass this cookie back in subsequent requests.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /login:
    post:
      summary: "Log in a user"
      operationId: loginUser
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginPayload"
            examples:
              bob:
                value:
                  email: "ligma2@gmail.com"
                  password: "testtest"
      responses:
        "200":
          description: "User Info"
          headers:
            Set-Cookie:
              description: >
                Contains the session cookies named `sid` and `user_id`.
                Pass this cookie back in subsequent requests.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    sidCookie:
      type: apiKey
      in: cookie
      name: finch-sid
  schemas:
    SignupPayload:
      type: object
      required:
        - email
        - password
        - first_name
        - last_name
        - income
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        income:
          type: number
    LoginPayload:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      required:
        - email
        # password hash never returned.
        - location
        - first_name
        - last_name
      properties:
        email:
          type: string
          format: email
        location:
          type: object
          required:
            - lat
            - lon
          properties:
            lat:
              type: number
            lon:
              type: number
        first_name:
          type: string
        last_name:
          type: string
        income:
          type: number
        # recurrings: # not returned
        # snapshots: # not returned
        # accounts: # maybe
        # goals: #
        # plans: #
        # insights: #
        device_url:
          type: string
          format: url
...
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string